requirements.txt
setuptools
dataclasses_json
aiofiles
pyyaml